Move env var collection impl into common
Part of #101857
